inverted logic to preserve file attributes (mod/access time and file permissions) by default
